No need for everyone who uses key constants to get a separate copy
