+++ ds3-sdk/src/main/java/com/spectralogic/ds3client/metadata/AbstractMetadataRestore.java
@@ -30,10 +30,10 @@ import java.util.concurrent.TimeUnit;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import static com.spectralogic.ds3client.metadata.interfaces.MetadataKeyConstants.KEY_ACCESS_TIME;
-import static com.spectralogic.ds3client.metadata.interfaces.MetadataKeyConstants.KEY_CREATION_TIME;
-import static com.spectralogic.ds3client.metadata.interfaces.MetadataKeyConstants.KEY_LAST_MODIFIED_TIME;
-import static com.spectralogic.ds3client.metadata.interfaces.MetadataKeyConstants.KEY_OS;
+import static com.spectralogic.ds3client.metadata.MetadataKeyConstants.KEY_ACCESS_TIME;
+import static com.spectralogic.ds3client.metadata.MetadataKeyConstants.KEY_CREATION_TIME;
+import static com.spectralogic.ds3client.metadata.MetadataKeyConstants.KEY_LAST_MODIFIED_TIME;
+import static com.spectralogic.ds3client.metadata.MetadataKeyConstants.KEY_OS;
 
 abstract class AbstractMetadataRestore implements MetadataRestore {
     private static final Logger LOG = LoggerFactory.getLogger(AbstractMetadataRestore.class);
