+++ ds3-sdk/src/main/java/com/spectralogic/ds3client/metadata/PosixMetadataRestore.java
@@ -29,13 +29,12 @@ import java.util.regex.Pattern;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import static com.spectralogic.ds3client.metadata.interfaces.MetadataKeyConstants.KEY_GID;
-import static com.spectralogic.ds3client.metadata.interfaces.MetadataKeyConstants.KEY_PERMISSION;
-import static com.spectralogic.ds3client.metadata.interfaces.MetadataKeyConstants.KEY_UID;
+import static com.spectralogic.ds3client.metadata.MetadataKeyConstants.KEY_GID;
+import static com.spectralogic.ds3client.metadata.MetadataKeyConstants.KEY_PERMISSION;
+import static com.spectralogic.ds3client.metadata.MetadataKeyConstants.KEY_UID;
 
 class PosixMetadataRestore extends AbstractMetadataRestore {
     private static final Logger LOG =LoggerFactory.getLogger(PosixMetadataRestore.class);
-
 
     public PosixMetadataRestore(final Metadata metadata, final String filePath, final String localOS, final MetadataRestoreListener metadataRestoreListener) {
         this.metadata = metadata;
