+++ ds3-sdk/src/main/java/com/spectralogic/ds3client/metadata/WindowsMetadataStore.java
@@ -33,10 +33,14 @@ import java.nio.file.attribute.AclFileAttributeView;
 import java.nio.file.attribute.BasicFileAttributes;
 import java.util.*;
 
-import static com.spectralogic.ds3client.metadata.interfaces.MetadataKeyConstants.*;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import static com.spectralogic.ds3client.metadata.MetadataKeyConstants.KEY_DACL;
+import static com.spectralogic.ds3client.metadata.MetadataKeyConstants.KEY_FLAGS;
+import static com.spectralogic.ds3client.metadata.MetadataKeyConstants.KEY_GROUP;
+import static com.spectralogic.ds3client.metadata.MetadataKeyConstants.KEY_OWNER;
+import static com.spectralogic.ds3client.metadata.MetadataKeyConstants.METADATA_PREFIX;
 
 class WindowsMetadataStore extends AbstractMetadataStore {
     private static final Logger LOG = LoggerFactory.getLogger(WindowsMetadataStore.class);
