+++ ds3-sdk/src/test/java/com/spectralogic/ds3client/metadata/PosixMetadataStore_Test.java
@@ -17,7 +17,6 @@ package com.spectralogic.ds3client.metadata;
 
 import com.google.common.collect.ImmutableMap;
 import com.spectralogic.ds3client.metadata.interfaces.MetadataStoreListener;
-import com.spectralogic.ds3client.metadata.interfaces.MetadataKeyConstants;
 import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
