+++ ds3-sdk/src/test/java/com/spectralogic/ds3client/metadata/WindowsMetadataStore_Test.java
@@ -16,7 +16,6 @@ package com.spectralogic.ds3client.metadata;
 
 import com.google.common.collect.ImmutableMap;
 import com.spectralogic.ds3client.metadata.interfaces.MetadataStoreListener;
-import com.spectralogic.ds3client.metadata.interfaces.MetadataKeyConstants;
 import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -30,7 +29,9 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.attribute.BasicFileAttributes;
 
-import static com.spectralogic.ds3client.metadata.interfaces.MetadataKeyConstants.*;
+import static com.spectralogic.ds3client.metadata.MetadataKeyConstants.KEY_GROUP;
+import static com.spectralogic.ds3client.metadata.MetadataKeyConstants.KEY_OWNER;
+import static com.spectralogic.ds3client.metadata.MetadataKeyConstants.METADATA_PREFIX;
 
 @RunWith(JUnit4.class)
 public class WindowsMetadataStore_Test {
@@ -43,7 +44,7 @@ public class WindowsMetadataStore_Test {
         final BasicFileAttributes attr = Files.readAttributes(file.toPath(), BasicFileAttributes.class);
         final long creationTime = attr.creationTime().toMillis();
         windowsMetadataStore.saveCreationTimeMetaData(attr);
-        Assert.assertEquals(mMetadataMap.build().get(MetadataKeyConstants.METADATA_PREFIX + MetadataKeyConstants.KEY_CREATION_TIME), String.valueOf(creationTime));
+        Assert.assertEquals(mMetadataMap.build().get(METADATA_PREFIX + MetadataKeyConstants.KEY_CREATION_TIME), String.valueOf(creationTime));
     }
 
     @Test
@@ -51,7 +52,7 @@ public class WindowsMetadataStore_Test {
         final BasicFileAttributes attr = Files.readAttributes(file.toPath(), BasicFileAttributes.class);
         final long lastAccessTime = attr.lastAccessTime().toMillis();
         windowsMetadataStore.saveAccessTimeMetaData(attr);
-        Assert.assertEquals(mMetadataMap.build().get(MetadataKeyConstants.METADATA_PREFIX + MetadataKeyConstants.KEY_ACCESS_TIME), String.valueOf(lastAccessTime));
+        Assert.assertEquals(mMetadataMap.build().get(METADATA_PREFIX + MetadataKeyConstants.KEY_ACCESS_TIME), String.valueOf(lastAccessTime));
     }
 
 
@@ -60,14 +61,14 @@ public class WindowsMetadataStore_Test {
         final BasicFileAttributes attr = Files.readAttributes(file.toPath(), BasicFileAttributes.class);
         final long lastModifiedTime = attr.lastModifiedTime().toMillis();
         windowsMetadataStore.saveLastModifiedTime(attr);
-        Assert.assertEquals(mMetadataMap.build().get(MetadataKeyConstants.METADATA_PREFIX + MetadataKeyConstants.KEY_LAST_MODIFIED_TIME), String.valueOf(lastModifiedTime));
+        Assert.assertEquals(mMetadataMap.build().get(METADATA_PREFIX + MetadataKeyConstants.KEY_LAST_MODIFIED_TIME), String.valueOf(lastModifiedTime));
     }
 
     @Test
     public void saveOSMetaData_Test() throws IOException {
         final String localOSName = MetaDataUtil.getOS();
         windowsMetadataStore.saveOSMetaData(localOSName);
-        Assert.assertEquals(mMetadataMap.build().get(MetadataKeyConstants.METADATA_PREFIX + MetadataKeyConstants.KEY_OS), String.valueOf(localOSName));
+        Assert.assertEquals(mMetadataMap.build().get(METADATA_PREFIX + MetadataKeyConstants.KEY_OS), String.valueOf(localOSName));
     }
 
     @Test
